# Conflicts:
#	src/benchmark/benchmarks.cpp
#	src/bin/bench.cpp
#	src/bin/graph_view.cpp
#	src/graph/WeightedGraph.cpp
#	src/graph/WeightedGraph.h
#	src/reference/Astar.h
#	src/reference/BoostAStarSearch.h
#	src/ripple/Collision.h
#	src/ripple/RippleSearch.cpp
#	src/ripple/RippleThread.cpp
#	src/ripple/RippleThread.h

+++ src/graph/Map.h
@@ -73,7 +73,7 @@ public:
   float distance(Node a, Node b) const;
   static float distance(Point a, Point b);
 
-  double cost(Node from, Node to);
+  float cost(Node from, Node to);
 
   MapNeighbours neighbours(Node i);
 };
